project/admin.py file
Over here , we need to register all the classes we created under the models.py page. Actually, this class gets saved within a variable where then it is rendered on the html page, but side by side this procedure is also vital so as to keep your work safe, as while compiling the code if you are using an older system or a laptop with very low processing speed, or an outdated pip ... it is probable that full code might not be compiled, django files also give issues if we try to edit them sometimes.